requirements.txt
aiofiles==24.1.0
schedule==1.2.2
beautifulsoup4==4.13.4
googlesearch_python==1.3.0
loguru==0.7.3
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
Telethon==1.40.0
tqdm==4.67.1
transformers==4.52.2